testsuite: introduce 'default_gdb_start_post_cmd' as an overridable function
Introduce a function named 'default_gdb_start_post_cmd' that is
invoked at the end of 'default_gdb_start'.  The function is by default
empty and can be overridden in a board file or a test to perform
setup right after GDB starts.

The need for this function arose from the fact that doing
'maint set target-non-stop on' was not possible when using the
native-extended-gdbserver board file.  This board file overrides
gdb_start as below:

  proc gdb_start { } {
      # Spawn GDB.
      default_gdb_start

      # And then GDBserver, ready for extended-remote mode.
      gdbserver_start_multi

      return 0
  }

That is, GDB is started and an extended-remote target is defined
immediately following it.  However, to set the remote target to
non-stop mode, we would have to issue the 'maint set target-non-stop on'
code ...

  proc gdb_start { } {
      # Spawn GDB.
      default_gdb_start

-->   ### ... right here.

      # And then GDBserver, ready for extended-remote mode.
      gdbserver_start_multi

      return 0
  }

If the target-non-stop mode is set after making the connection,
we see the following problem:

   $ gdb
   (gdb) file a.out
   ...
   (gdb) set remote exec-file a.out
   (gdb) target extended-remote | gdbserver --once --multi -
   ...
   (gdb) maint set target-non-stop on
   (gdb) run
   ...
   Unexpected vCont reply in non-stop mode: T05swbreak:;06:f0dfffffff7f0000;07:10deffffff7f0000;10:d194ddf7ff7f0000;thread:p4706.4706;core:d;

This patch provides a generalized solution and gives the ability to do
custom setup after starting GDB but before connecting to the target.

gdb/testsuite/ChangeLog:
2020-04-20  Tankut Baris Aktemur  <tankut.baris.aktemur@intel.com>

	* lib/gdb.exp (default_gdb_start_post_cmd): New overridable
	function.
	(default_gdb_start): Invoke 'default_gdb_start_post_cmd'
	at the end.