gdb/jit: enable tracking multiple jitter objfiles
GDB's JIT handler stores an objfile (and data associated with it) per
program space to keep track of JIT breakpoint information.  This
assumes that there is at most one JITer objfile in the program space.
However, there may be multiple.  If so, only the first JITer's hook
breakpoints would be realized and the JIT events from the other JITers
would be missed.  This patch replaces the single objfile pointer with
a list so that multiple objfiles can be tracked per program space.

This patch is viewed better with 'git diff -w'.

Regression-tested on X86_64 Linux.

gdb/ChangeLog:
2020-06-15  Tankut Baris Aktemur  <tankut.baris.aktemur@intel.com>

	* jit.c: (struct jiter_and_bp): New struct.
	(struct jit_program_space_data): Enable storing multiple objfiles.
	(jit_breakpoint_re_set_internal): Iterate over all objfiles in the
	program space to look for JIT symbols.
	(jit_breakpoint_deleted): Update to iterate over multiple jiters.
	(free_objfile_data): Ditto.
	(jit_inferior_init): Ditto.

gdb/testsuite/ChangeLog:
2020-06-15  Tankut Baris Aktemur  <tankut.baris.aktemur@intel.com>

	* gdb.base/jit-reader-simple.exp: Add a scenario for a binary that
	loads two JITers.