gdb/regcache: return REG_UNAVAILABLE in raw_read if NOT_AVAILABLE_ERROR is seen
In regcache's raw_read, it is possible that 'raw_update' fails with an
exception.  This could happen, for instance, if the debugged thread is
already exited and the target cannot fetch its registers.  Catch the
exception and return REG_UNAVAILABLE if the error is of kind
NOT_AVAILABLE_ERROR.  This makes clients' lives easier.

Regression-tested on X86_64-Linux.